target/ppc/spapr: Override H_GET_CPU_CHARACTERISTICS with environment variables
Override the response formulated in the hypercall handler with values
of the H_CPU_CHARS and H_CPU_BEHAV environment variables, if defined.

Optionally, allow the values used in the first call (i.e., boot time)
to be different, with the H_CPU_CHARS_1 and H_CPU_BEHAV_1 environment
variables, if defined.

This is useful to test and debug the kernel-side implementation of
the various possible permutations without relying on real firmware
and hardware revisions that provide the desired settings.

Keep all that under DEBUG_H_GET_CPU_CHARACTERISTICS define guards.

Signed-off-by: Mauricio Faria de Oliveira <mauricfo@linux.vnet.ibm.com>